lently-backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
google-generativeai==0.3.2
google-auth==2.27.0
google-cloud-pubsub==2.19.0
google-cloud-secret-manager==2.18.0
google-api-python-client==2.116.0
firebase-admin==6.4.0
redis==5.0.1
httpx==0.26.0
python-jose[cryptography]==3.3.0
slowapi==0.1.9
pytest==7.4.4
pytest-asyncio==0.23.3
